Lronac2isis tests added (#4141)
* added new files

* added lronac2isis tests

* added data (important)

* static funcs
